[Refactored] add otp_text_field barrier file
