+++ example/lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:otp_text_field/otp_field.dart';
+
 import 'package:otp_text_field/otp_text_field.dart';
-import 'package:otp_text_field/style.dart';
+
 
 void main() => runApp(MyApp());
